Add fifth and sixth sections
