+++ src/assets/data/marketplaceSections.js
@@ -94,4 +94,70 @@ const fourthMarketplace = [{
   price: '$25,000',
 }];
 
-export { firstMarketplace, thirdMarketplace, fourthMarketplace };
+const fifthMarketplace = [{
+  followers: '../multimedia/img/icons/marketplace/followers/100k.png',
+  platformLogo: '../multimedia/img/icons/social/light/__tiktok-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/5m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__youtube-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10k.png',
+  platformLogo: '../multimedia/img/icons/social/light/__instagram-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__twitter-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__snap-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__fb-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}];
+
+const sixthMarketplace = [{
+  followers: '../multimedia/img/icons/marketplace/followers/100k.png',
+  platformLogo: '../multimedia/img/icons/social/light/__tiktok-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/5m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__youtube-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10k.png',
+  platformLogo: '../multimedia/img/icons/social/light/__instagram-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__twitter-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__snap-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}, {
+  followers: '../multimedia/img/icons/marketplace/followers/10m.png',
+  platformLogo: '../multimedia/img/icons/social/light/__fb-logo.png',
+  companyLogo: '../multimedia/img/icons/marketplace/yourlogohere.png',
+  price: '$25,000',
+}];
+
+export {
+  firstMarketplace, thirdMarketplace, fourthMarketplace, fifthMarketplace, sixthMarketplace,
+};
